Add check that shouldn't need to be checked
Somehow on tests, the temp folder would render itself unfindable, and thus made the program crash.

+++ DATFromDir/DATFromDir.cs
@@ -95,7 +95,10 @@ namespace SabreTools
 				}
 
 				// Delete the temp directory
-				Directory.Delete(_tempDir, true);
+				if (Directory.Exists(_tempDir))
+				{
+					Directory.Delete(_tempDir, true);
+				}
 			}
 
 			// Order the roms by name of parent, then name of rom
